refactor(gui): manage nested users tab route based
